+++ Mage/src/main/java/mage/abilities/dynamicvalue/common/ManacostVariableValue.java
@@ -16,7 +16,7 @@ public enum ManacostVariableValue implements DynamicValue {
 
     @Override
     public int calculate(Game game, Ability sourceAbility, Effect effect) {
-        return (int) CardUtil.getSourceCostsTag(game, sourceAbility, "X", 0);
+        return CardUtil.getSourceCostsTag(game, sourceAbility, "X", 0);
     }
 
     @Override
